Update MIN_HANDSHAKE_WAIT_TIME to 85 microseconds
Aligned MIN_HANDSHAKE_WAIT_TIME with the Amiga 500 Technical Reference Manual's recommended value of 85 microseconds. Previously, it was set to 65 microseconds, which worked for my specific A500 keyboard model but may not be compatible with all keyboard variations and revisions. The value can be increased further if instability is observed, without noticeable delays in keyboard performance.